requirements.txt
Pillow
bitarray
